+++ src/ldc/filter/__init__.py
@@ -10,6 +10,7 @@ from ._record_window import RecordWindow
 from ._remove_blocks import RemoveBlocks
 from ._remove_empty import RemoveEmpty
 from ._remove_patterns import RemovePatterns
+from ._replace_patterns import ReplacePatterns
 from ._reset_ids import ResetIDs
 from ._skip_duplicate_ids import SkipDuplicateIDs
 from ._skip_duplicate_text import SkipDuplicateText